Create OOP CT1
Write the class named “Car” to represent a car in a 2-D coordinate
system. Complete the following task sequentially:
 Create a class named “Point” to represent a 2-D point with x & y.
 Add constructor to the “Point” class to set value of x & y.
 Add a field named “position” with type “Point” in the “Car” class.
 Add a constructor to the “Car” class that takes the initial position of
the car as argument and sets the “Position” field properly within the
constructor.
 Add a method named “move_forward” in the class “Car” that takes
one parameter: “amount”.
The method will increment the value of x or y of the position by the
value of “amount”.
Assume, all fields are of double datatype.
Write the main method with the Main Class to initialize an array of
class named “Car” with size of 10. Allocate objects for all the indices
in the array using a loop