Add files via upload
You have been hired by a software development company to create a program that simulates a zoo. The program should allow users to manage different types of animals in the zoo, such as mammals, birds, and reptiles. Your task is to design and implement the necessary classes to represent the animals and their methods.

Design a base class called Animal that has the following
attributes:
name (string): to store the name of the animal.
age (integer): to store the age of the animal
Encapsulate these attributes using appropriate access modifiers to ensure data security.

Implement the following methods in the Animal class:
eat(): void: This method should display a message indicating that the animal is eating.
sleep(): void: This method should display a message indicating that the animal is sleeping.

Create three derived classes:
Mammal, Bird, and Reptile, which inherit from the Animal class.

Each derived class should have specific attributes and methods:
Mammal:
Attribute: furColor (string): to store the color of the mammal's fur
Method: Override the eat() method to display a message indicating that the mammal is eating its favorite food.
Bird:
Attribute: wingspan (float): to store the wingspan of the bird
[10]
Method: Overload the sleep() method with a parameterized version that takes an additional argument representing the number of hours the bird will sleep.
Reptile:
Attribute: scaleColor (string): to store the color of the reptile's scales
Method: Implement a new method called crawl(): void that displays a message indicating that the reptile is crawling.
Create instances of each derived class Inside the main method (Main Class) to represent different animals in the zoo. Set the values of the attributes accordingly.
Demonstrate method overloading by calling the sleep() method on a bird instance with different arguments (e.g., 6 hours, 8 hours) and observing the different sleep durations. Demonstrate method overriding by calling the eat() method on a mammal instance and observing the overridden behavior. Call the crawl() method on a reptile instance to demonstrate the unique behavior implemented in the Reptile class.